+++ docs/changelog.rst
@@ -12,6 +12,7 @@ New Features
 ============
 - Add no_labels option to waves visualize (:issue:`583`, :merge:`717`). By `Prabhu Khalsa`_.
 - New print-tree feature in waves visualize (:issue:`582`, :merge:`718`). By `Prabhu Khalsa`_.
+- Add input-file option to waves visualize to graph provided input (:issue:`584`, :merge:`719`). By `Prabhu Khalsa`_.
 
 Bug fixes
 =========